Improved main_tick handling and cleaned up pktrcv
All of the main loop delays are tied to 'main_tick' so I added a
new 'tick_count' which allows for a repeated main_tick value.
Also added needed documentation on how these work. In addition,
cleaned up the receive code to be a bit more efficient and modified
the state processing to not touch tick_count unless necessary.