Extensive changes to make Bidir comms work
When sending a packet via the controller, with a STATUS_REQUEST type
(and eventually including EEPROM_READ), the system switches to RX
mode after the transmit. If a packet is received within the next
four passes, it is sent upstairs. Otherwise, it reverts back to
transmitting.

Right now this is polled and it's not reliable. Ideally it would
use the RX-IRQ. Also, it takes a long time to get a response. It
could be worth optimising this as not much can happen during that
window.